backend/requirements.txt
tweepy
textblob
python-dotenv
psycopg2-binary